Persist session data outside of Stub IdP
Currently Stub IdP stores all session data in an in-memory cache. When the application stops as part of a deployment, all session data currently in-memory is lost, meaning that anyone using the Stub IdP at this time will need to restart their entire Verify journey.

This commit refactors the Stub IdP to be able to serialize/deserialize session data to and from an external data store. It is intended as a solution to tackle the immediate issue of zero-downtime deploys of Stub IdP; refactoring the stored session data to remove unnecessary dependencies may help to remove some of the brittleness around serialization and deserialization, especially of external types, is a piece of work that will come later on.

Co-authored-by: Rosa Fox <rosa.fox@digital.cabinet-office.gov.uk>