EID-841: Fix gradle and scripts for PaaS deploy
Now that stub-idp is a subproject within this repository, we need to
update the paths and filenames in the various gradle build files and
shell scripts accordingly. Also, this adds a `jenkins.sh` script to
orchestrate the Jenkins build, so that we don't have to constantly edit
the job manually.