BAU: Separate Countries and IDPs in config to fix non-prod Stub IDP journeys
The stub-idp.yml config file now includes stub-country so we can access it via the API to query and create users. However, this means that it's treated as an IDP and therefore eIDAS journeys on non-prod environments are not working. This change makes sure the code treats countries and IDPs separately and that it doesn't create eIDAS users in IDP format, which is what made Stub IDP broken.

Summary of changes:
  - Add a class to easily access the stub-idps.yml config file which contains the credentials for all stub IDPs and stub-country
  - Separate credentials handling into IdpStubsRepository for IDPs and StubCountryRepository for countries
  - Make the auth filter aware of the difference
  - Make the new class injectable

Fixed a failing test following the library update