De-optionalify IdentityProviderAuthnStatementBuilder
We always reference userIpAddress with `orElse(null)` so we can just have it be
a nullable variable and drop the Optional

We only wrap FraudAuthnDetails in an Optional so we can decide what to do if
it's present. We can do the same thing with a check against null, so remove the
Optional there too.