Stop publishing to artifactory
Bintray is the source of truth now, for stub-idp-saml and
stub-idp-saml-test.

Bintray is made available via whitelisted-repos on artifactory, so we
still pull dependencies down using artifactory, but we only need to
push dependencies to bintray.

The exception is stub-idp itself; our artifact store for this is S3;
concourse builds stub-idp and uploads the artifacts to an S3 bucket,
and deploys from there to PaaS.