HUB-770: Config for running a new G3 stub-idp
This has been pushed to the PaaS integration stubs space and is running.
Private keys have been redacted from the manifest.yml but can be fetched
from the env of the running application.

The private key to use for the signing cert is in blackbox

To deploy this version to integration:

* Update manifest.yml with the private keys -
  `STUB_IDP_SIGNING_PRIVATE_KEY` and `STUB_COUNTRY_SIGNING_PRIVATE_KEY`
* Log in to PaaS with cf cli - you need the govuk-verify org and the
  integration-stubs space
* run `cf push stub-idp-integration-g3 -f manifest.yml -p stub-idp/build/distributions/stub-idp-unspecified.zip --strategy rolling`