HUB-915: Pull from maven central
Update dockerfile to use version 6.8 of gradle - the vendored version
has shifted to v7 which doesn't work.

Remove the versions plugin, and the buildscript block which was only
there to bring it in.

Bump versions of dependencies to ensure they're being pulled from maven
central.

Update public builds to use maven central and remove redundant repos.

Update `SamlDecrypter` to work with the new versions of hub-saml.