BAU: Fix distZip artifact name in pipeline
A change was made a while ago so that version used by distZip for naming
was set to the value of `build_version` in the root gradle file.
`build_version` is composed from the open saml version set as well as
the env variable `BUILD_NUMBER`. If `BUILD_NUMBER` is not set then
`SNAPSHOT` is used instead.

This pipeline hasn't been updated to reflect this and so all new
artifacts have the same name - `stub-idp-3.4.3-SNAPSHOT.zip`

This all means that the pipeline doesn't register a new build artifact
as a new version and so it doesn't get released.

Providing the Concourse version number as BUILD_NUMBER ensures each new
build has a unique artifact name and triggers the pipeline correctly.